requirements.txt
beautifulsoup4==4.11.2
Flask==2.2.3
requests==2.28.2
python-dotenv==1.0.0
openai==1.75.0
google-api-python-client==2.109.0
werkzeug==2.3.0